work on multiple source support
